Move sampling_rate back to info.py
